Initial trimming of ScreenSpacePositions
