+++ Map Map/Background Map/Background Map Layers/BackgroundMapButtonsV.swift
@@ -16,8 +16,6 @@ struct BackgroundMapButtonsV: View {
     @FetchRequest(sortDescriptors: []) private var markers: FetchedResults<Marker>
     /// Current Core Data managed object context.
     @Environment(\.managedObjectContext) private var moc
-    /// Screen-space positions of Markers, MapMaps, and the user's location.
-    @Environment(ScreenSpacePositionsM.self) private var screenSpacePositions
     /// Details about the background map.
     @Environment(BackgroundMapDetailsM.self) private var backgroundMapDetails
     /// Tracker for adding or removing markers.
@@ -28,6 +26,8 @@ struct BackgroundMapButtonsV: View {
     let screenSize: CGSize
     /// Background map ID.
     let mapScope: Namespace.ID
+    
+    let mapContext: MapProxy
     
     /// Type for tracking adding or removing markers.
     enum MarkerButtonType: Equatable {
@@ -98,7 +98,7 @@ struct BackgroundMapButtonsV: View {
     
     func checkOverMarker() {
         for marker in markers {
-            if let markerPos = screenSpacePositions[marker] {
+            if let markerPos = mapContext.convert(marker.coordinates, to: .global) {
                 let xComponent = abs(markerPos.x - screenSize.width / 2)
                 let yComponent = abs(markerPos.y - screenSize.height / 2)
                 let distance = sqrt(pow(xComponent, 2) + pow(yComponent, 2))
@@ -120,11 +120,8 @@ struct BackgroundMapButtonsV: View {
         let newMarker = Marker(coordinates: backgroundMapDetails.position, insertInto: moc)
         let centerPoint: CGPoint = CGPoint(size: screenSize / 2)
         for mapMap in mapMaps {
-            if let path = screenSpacePositions.generateMapMapRotatedBounds(
-                mapMap: mapMap,
-                backgroundMapRotation: backgroundMapDetails.rotation
-            )?.cgPath, 
-                path.contains(centerPoint) {
+            if let path = BackgroundMap.generateMapMapRotatedConvexHull(mapMap: mapMap, backgroundMapDetails: backgroundMapDetails, mapContext: mapContext)?.cgPath,
+               path.contains(centerPoint) {
                 newMarker.addToMapMap(mapMap)
             }
         }
