+++ Map Map/Background Map/Background Map Layers/BackgroundMapV.swift
@@ -16,10 +16,6 @@ struct BackgroundMap: View {
     @FetchRequest(sortDescriptors: []) private var markers: FetchedResults<Marker>
     /// Background map details to be updated by this map.
     @Environment(BackgroundMapDetailsM.self) private var backgroundMapDetails: BackgroundMapDetailsM
-    /// GPS user location.
-    @State private var locationsHandler = LocationsHandler.shared
-    /// Screen-space positions of MapMaps, Markers, and the user location.
-    @Environment(ScreenSpacePositionsM.self) private var screenSpacePositions
     /// Track if MapMaps are tappable.
     @State private var tappableMapMaps = true
     /// Background map ID.
@@ -69,51 +65,60 @@ struct BackgroundMap: View {
             backgroundMapDetails.position = update.region.center
             backgroundMapDetails.span = update.region.span
             MKMetersPerMapPointAtLatitude(update.region.center.latitude)
-            if let screenSpaceUserLocation = mapContext.convert(locationsHandler.lastLocation.coordinate, to: .local) {
-                self.screenSpacePositions.userLocation = screenSpaceUserLocation
-            }
-            var screenSpaceMarkerPositions: [Marker : CGPoint] = [:]
-            for marker in markers {
-                guard let screenSpaceMarkerPosition = mapContext.convert(marker.coordinates, to: .local)
-                else { return }
-                screenSpaceMarkerPositions[marker] = screenSpaceMarkerPosition
-            }
-            self.screenSpacePositions.setPositions(screenSpaceMarkerPositions)
-            var screenSpaceMapMapPositions: [MapMap : CGRect] = [:]
-            for mapMap in mapMaps {
-                guard let center = mapContext.convert(mapMap.coordinates, to: .local)
-                else { return }
-                let size = backgroundMapDetails.scale * mapMap.mapMapScale
-                let rect = CGRect(
-                    origin: center,
-                    size: CGSize(
-                        width: size,
-                        height: size / mapMap.imageWidth * mapMap.imageHeight
-                    )
-                )
-                screenSpaceMapMapPositions[mapMap] = rect
-            }
-            self.screenSpacePositions.setPositions(screenSpaceMapMapPositions)
         }
         .mapStyle(.standard(elevation: .realistic))
-        .onAppear { locationsHandler.startLocationTracking() }
-        .onDisappear { locationsHandler.stopLocationTracking() }
-        .onChange(of: $locationsHandler.lastLocation.wrappedValue) { _, newLocation in
-            if let screenSpaceUserLocation = mapContext.convert(newLocation.coordinate, to: .local) {
-                self.screenSpacePositions.userLocation = screenSpaceUserLocation
-            }
-        }
-        .onReceive(NotificationCenter.default.publisher(for: .editedMarkerLocation)) { notification in
-            if let marker = notification.userInfo?["marker"] as? Marker,
-               let screenSpaceMarkerLocation = mapContext.convert(marker.coordinates, to: .local) {
-                screenSpacePositions[marker] = screenSpaceMarkerLocation
-            }
-        }
         .onReceive(NotificationCenter.default.publisher(for: .editingMarker)) { notification in
             if let editingStatus = notification.userInfo?["editing"] as? Bool {
                 tappableMapMaps = !editingStatus
             }
         }
-        .animation(.linear, value: screenSpacePositions.userLocation)
+    }
+    
+    /// Given a specific rotation and MapMap, a UIBezierPath is formed to allow for precise calculations.
+    /// - Parameters:
+    ///   - mapMap: MapMap to generate a rotated bounding box for.
+    ///   - backgroundMapRotation: Background map rotation.
+    /// - Returns: A bounding box of a MapMap that has been rotated to match the MapMap.
+    public static func generateMapMapRotatedConvexHull(mapMap: MapMap, backgroundMapDetails: BackgroundMapDetailsM, mapContext: MapProxy) -> UIBezierPath? {
+        guard let rect: CGRect = BackgroundMap.generateMapMapBounds(mapMap, mapContext: mapContext, backgroundMapScale: backgroundMapDetails.scale)
+        else { return nil }
+        let transform = CGAffineTransform(translationX: rect.midX - rect.width / 2, y: rect.midY - rect.height / 2)
+            .rotated(by: (backgroundMapDetails.rotation - Angle(degrees: mapMap.mapMapRotation)).radians)
+            .translatedBy(x: -rect.midX, y: -rect.midY)
+        let padding: CGFloat = 25
+        let rotatedPoints: [CGPoint] = // Calculate rotated points
+        [
+            CGPoint(x: rect.minX - padding, y: rect.minY - padding).applying(transform),
+            CGPoint(x: rect.maxX + padding, y: rect.minY - padding).applying(transform),
+            CGPoint(x: rect.maxX + padding, y: rect.maxY + padding).applying(transform),
+            CGPoint(x: rect.minX - padding, y: rect.maxY + padding).applying(transform)
+        ]
+        // Check if the point is within the convex hull of rotated points
+        let convexHullPath = UIBezierPath()
+        convexHullPath.move(to: rotatedPoints[0])
+        convexHullPath.addLine(to: rotatedPoints[1])
+        convexHullPath.addLine(to: rotatedPoints[2])
+        convexHullPath.addLine(to: rotatedPoints[3])
+        convexHullPath.close()
+        return convexHullPath
+    }
+    
+    /// Creates a bounding box for a given MapMap.
+    /// - Parameters:
+    ///   - mapMap: MapMap to generate the bounding box for.
+    ///   - mapContext: Map Proxy to use when generating the given MapMap's screenspace location.
+    ///   - backgroundMapScale: Scale of the background map.
+    /// - Returns: Bounds of the MapMap.
+    public static func generateMapMapBounds(_ mapMap: MapMap, mapContext: MapProxy, backgroundMapScale: CGFloat) -> CGRect? {
+        guard let center = mapContext.convert(mapMap.coordinates, to: .local)
+        else { return nil }
+        let size = backgroundMapScale * mapMap.mapMapScale
+        return CGRect(
+            origin: center,
+            size: CGSize(
+                width: size,
+                height: size / mapMap.imageWidth * mapMap.imageHeight
+            )
+        )
     }
 }
