+++ Map Map/MapMap Editing/MapMapEditorV.swift
@@ -6,6 +6,7 @@
 //
 
 import Bottom_Drawer
+import MapKit
 import SwiftUI
 
 /// One-stop-shop for editing MapMaps.
@@ -16,14 +17,16 @@ struct MapMapEditor: View {
     @Environment(BackgroundMapDetailsM.self) private var backgroundMapDetails
     /// Current Managed Object Context.
     @Environment(\.managedObjectContext) private var moc
-    /// All screen space positions for MapMaps, Markers, and the user location.
-    @Environment(ScreenSpacePositionsM.self) private var screenSpacePositions
     /// Desired name for the current MapMap.
     @State private var workingName: String = ""
     /// Placement of the MapMap.
     @State private var mapMapPosition: CGRect = .zero
     /// Tracker for showing the photo (not MapMap) editor.
     @State private var showingPhotoEditor = false
+    
+    @FetchRequest(sortDescriptors: []) var markers: FetchedResults<Marker>
+    
+    let mapContext: MapProxy
     
     var body: some View {
         ZStack {
@@ -72,9 +75,7 @@ struct MapMapEditor: View {
                             mapMap.mapDistance = 1 / backgroundMapDetails.scale
                             mapMap.isEditing = false
                             mapMap.isSetup = true
-                            screenSpacePositions[mapMap] = mapMapPosition
-                            if let overlappingMarkers = 
-                                screenSpacePositions.mapMapOverMarkers(mapMap, backgroundMapRotation: backgroundMapDetails.rotation ) {
+                            if let overlappingMarkers = MapMapEditor.mapMapOverMarkers(mapMap, backgroundMapDetails: backgroundMapDetails, mapContext: mapContext) {
                                 for marker in mapMap.formattedMarkers { mapMap.removeFromMarkers(marker) } // Remove MapMap from all Markers
                                 for marker in overlappingMarkers { mapMap.addToMarkers(marker) } // Add MapMap to all
                             }
@@ -108,5 +109,27 @@ struct MapMapEditor: View {
         .fullScreenCover(isPresented: $showingPhotoEditor, content: {
             PhotoEditorV(mapMap: mapMap)
         })
+    }
+    
+    /// Determine all Markers that overlap a given MapMap
+    /// - Parameters:
+    ///   - mapMap: MapMap to check Marker positions against.
+    ///   - backgroundMapRotation: Background map rotation.
+    /// - Returns: If the MapMap's position in screen-space was not found, then nil is returned.
+    /// Otherwise, all available Markers are returned.
+    public static func mapMapOverMarkers(_ mapMap: MapMap, backgroundMapDetails: BackgroundMapDetailsM, mapContext: MapProxy) -> [Marker]? {
+        guard let mapMapBounds = BackgroundMap.generateMapMapRotatedConvexHull(mapMap: mapMap, backgroundMapDetails: backgroundMapDetails, mapContext: mapContext)?.cgPath
+        else { return nil }
+        var markers: [Marker] = []
+        
+        for marker in markers {
+            if let markerPosition = mapContext.convert(marker.coordinates, to: .global) {
+                let markerBounds = MarkerEditorV.generateMarkerBoundingBox(markerPosition: markerPosition)
+                if mapMapBounds.intersects(markerBounds) {
+                    markers.append(marker)
+                }
+            }
+        }
+        return markers
     }
 }
