+++ Map Map/Marker Editor/MarkerEditorV.swift
@@ -6,6 +6,7 @@
 //
 
 import Bottom_Drawer
+import MapKit
 import SwiftUI
 
 /// A one-stop-shop for editing details about a Marker.
@@ -18,18 +19,21 @@ struct MarkerEditorV: View {
     @State private var showingImagePicker: Bool = false
     /// Current Core Data managed object context.
     @Environment(\.managedObjectContext) private var moc
-    /// Screen space positions for plotted objects.
-    @Environment(ScreenSpacePositionsM.self) private var screenSpacePositions
     /// Marker being edited.
     @ObservedObject var marker: FetchedResults<Marker>.Element
     /// Information about the background map being plotted on top of.
     @Environment(BackgroundMapDetailsM.self) private var backgroundMapDetails
     
-    init(marker: FetchedResults<Marker>.Element) {
+    @FetchRequest(sortDescriptors: []) var mapMaps: FetchedResults<MapMap>
+    
+    let mapContext: MapProxy
+    
+    init(marker: FetchedResults<Marker>.Element, mapContext: MapProxy) {
         self.marker = marker
         if let name = marker.name { self._workingName = State(initialValue: name) }
         else { self._workingName = State(initialValue: "") }
         self._saveAngle = State(initialValue: marker.lockRotationAngleDouble != nil)
+        self.mapContext = mapContext
         NotificationCenter.default.post(name: .editingMarker, object: nil, userInfo: ["editing":true])
     }
     
@@ -87,7 +91,6 @@ struct MarkerEditorV: View {
                                 marker.name = workingName
                                 marker.coordinates = backgroundMapDetails.position
                                 marker.lockRotationAngleDouble = saveAngle ? backgroundMapDetails.rotation.degrees : nil
-                                determineMarkerSSLocation(geo: geo)
                                 determineMarkerOverlap()
                                 try? moc.save()
                             } label: {
@@ -118,8 +121,8 @@ struct MarkerEditorV: View {
     }
     
     func determineMarkerOverlap() {
-        if let overlappingMapMaps =
-            screenSpacePositions.markerOverMapMaps(marker, backgroundMapRotation: backgroundMapDetails.rotation) {
+        if let overlappingMapMaps = MarkerEditorV.markerOverMapMaps(marker, backgroundMapDetails: backgroundMapDetails, mapContext: mapContext, mapMaps: mapMaps)
+            /*screenSpacePositions.markerOverMapMaps(marker, backgroundMapRotation: backgroundMapDetails.rotation)*/ {
             // Remove current marker from all MapMaps
             for mapMap in marker.formattedMapMaps { mapMap.removeFromMarkers(marker) }
             // Add Marker to relevant MapMaps
@@ -127,10 +130,38 @@ struct MarkerEditorV: View {
         }
     }
     
-    func determineMarkerSSLocation(geo: GeometryProxy) {
-        screenSpacePositions[marker] = CGPoint(
-            x: geo.size.width / 2,
-            y: geo.size.height / 2 + geo.safeAreaInsets.top - 2
-        )
+    /// Determine all MapMaps that overlap a given Marker
+    /// - Parameters:
+    ///   - marker: Marker to check MapMap positions against.
+    ///   - backgroundMapRotation: Background map rotation.
+    /// - Returns: If the Marker's position in screen-space was not found, then nil is returned.
+    /// Otherwise, all available MapMaps are returned.
+    public static func markerOverMapMaps(_ marker: Marker, backgroundMapDetails: BackgroundMapDetailsM, mapContext: MapProxy, mapMaps: FetchedResults<MapMap>) -> [MapMap]? {
+        guard let markerPosition = mapContext.convert(marker.coordinates, to: .global)
+        else { return nil }
+        let marker = MarkerEditorV.generateMarkerBoundingBox(markerPosition: markerPosition)
+        var mapMaps: [MapMap] = []
+        for mapMap in mapMaps {
+            if let mapMapBounds = BackgroundMap.generateMapMapRotatedConvexHull(mapMap: mapMap, backgroundMapDetails: backgroundMapDetails, mapContext: mapContext)?.cgPath,
+                mapMapBounds.intersects(marker) {
+                mapMaps.append(mapMap)
+            }
+        }
+        return mapMaps
+    }
+    
+    /// Generate a Marker's bounding box
+    /// - Parameter markerPosition: Center point of the bounding box.
+    /// - Returns: A CGPath of a circle centered at the marker position.
+    public static func generateMarkerBoundingBox(markerPosition: CGPoint) -> CGPath {
+        return UIBezierPath(
+            ovalIn: CGRect(
+                origin: markerPosition,
+                size: CGSize(
+                    width: BackgroundMapPointsV.iconSize,
+                    height: BackgroundMapPointsV.iconSize
+                )
+            )
+        ).cgPath
     }
 }
